requirements.txt
python-binance==1.0.12
sqlalchemy==1.4.15
schedule==1.1.0
apprise==0.9.5.1
Flask==1.1.2
gunicorn==20.1.0
flask-cors==3.0.10
flask-socketio==5.0.1
eventlet==0.30.2
python-socketio[client]==5.2.1
cachetools==4.2.2
sqlitedict==1.7.0
unicorn-binance-websocket-api==1.34.2
unicorn-fy==0.11.0
